# sample/service-user-center/docs/api/single.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/role": {
      "post": {
        "operationId": "RoleHttpController_insert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleInsertReqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleInsertResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Role - 角色信息"
        ]
      },
      "get": {
        "operationId": "RoleHttpController_selectByPage",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isDefault",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleSelectByPageResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Role - 角色信息"
        ]
      }
    },
    "/api/role/{id}": {
      "get": {
        "operationId": "RoleHttpController_selectById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleSelectByIdResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Role - 角色信息"
        ]
      },
      "patch": {
        "operationId": "RoleHttpController_updateById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdateReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateByIdResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Role - 角色信息"
        ]
      },
      "delete": {
        "operationId": "RoleHttpController_deleteById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteByIdResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Role - 角色信息"
        ]
      }
    },
    "/api/role/m": {
      "post": {
        "operationId": "RoleHttpController_insertMulti",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleInsertMultiResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Role - 角色信息"
        ]
      }
    },
    "/api/role/m/{ids}": {
      "get": {
        "operationId": "RoleHttpController_selectByIds",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "path",
            "schema": {
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleSelectByIdsResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Role - 角色信息"
        ]
      },
      "patch": {
        "operationId": "RoleHttpController_updateByIds",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "path",
            "schema": {
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdateReqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateByIdsResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Role - 角色信息"
        ]
      },
      "delete": {
        "operationId": "RoleHttpController_deleteByIds",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "path",
            "schema": {
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteByIdsResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Role - 角色信息"
        ]
      }
    },
...
  },
  "info": {
    "title": "SAMPLE SERVICE USER CENTER",
    "description": "",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PageResDto": {
        "type": "object",
        "properties": {
          "pageSize": {
            "type": "number"
          },
          "pageNum": {
            "type": "number"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "pageSize",
          "pageNum",
          "total"
        ]
      },
...
      "UpdateEffectInfo": {
        "type": "object",
        "properties": {
          "effectNum": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "effectNum"
        ]
      },
      "UpdateByIdResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/UpdateEffectInfo"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "DeleteEffectInfo": {
        "type": "object",
        "properties": {
          "effectNum": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "effectNum"
        ]
      },
      "DeleteByIdResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/DeleteEffectInfo"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
...
      "UpdateByIdsResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/UpdateEffectInfo"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "DeleteByIdsResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/DeleteEffectInfo"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "RoleInsertReqDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "isDefault": {
            "type": "number"
          }
        }
      },
      "RoleInsertResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "isDefault": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "isDefault"
        ]
      },
      "RoleInsertResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/RoleInsertResDto"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "RoleSelectResDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "isDefault": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "isDefault"
        ]
      },
      "RoleSelectByPageResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleSelectResDto"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageResDto"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "page",
          "message"
        ]
      },
      "RoleSelectByIdResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RoleSelectResDto"
              }
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "RoleUpdateReqDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "isDefault": {
            "type": "number"
          }
        }
      },
      "RoleInsertMultiResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleInsertResDto"
            }
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
      "RoleSelectByIdsResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleSelectResDto"
            }
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "data",
          "message"
        ]
      },
...
    }
  }
}